+++ lib/flutter_reactive_n.dart
@@ -1,5 +1,5 @@
-
 part of 'flutter_reactive.dart';
+
 /// Nullable Reactive
 class ReactiveN<T> extends Reactive<T?> {
   ReactiveN([super._value, super._strict]);