+++ test/flutter_reactive_test.dart
@@ -8,9 +8,29 @@ void dp([dynamic v = '']) {
 }
 
 void main() {
+  test("Computed reactives", () {
+    final r1 = 0.reactive();
+    final r2 = 0.2.reactive();
+    final rC = Reactive.compute(() => r1.value + r2.value);
+    final rC2 = Reactive.combine2(r1, r2, (v1, v2) => v1 - v2);
+
+    rC.listen((c) {
+      dp(c);
+    });
+
+    rC2.listen((c) {
+      dp('c2 $c');
+    });
+
+    r1.increment();
+    r1.increment();
+    r2.increment(5);
+  });
   test('Validator', () {
     final counter = 0.reactive();
-    counter.require((v) => v > 0).require((v) => v < 4, "Te");
+    counter
+        .require((v) => v > 0)
+        .require((v) => v < 4, "Counter should be less than 4");
 
     counter.listen((v) {
       dp("Counter: $v");